Hilt configurated, injecting Retrofit and OkHttClient

+++ app/src/main/java/dev/brunofelix/marvelapp/data/remote/dto/CharacterDto.kt
@@ -5,7 +5,7 @@ import dev.brunofelix.marvelapp.domain.model.Character
 
 data class CharacterDto(
     @SerializedName("id")
-    var id: Long?,
+    var id: Int?,
 
     @SerializedName("name")
     var name: String?,
